+++ iod/src/clockwork.cpp
@@ -108,7 +108,8 @@ void usage(int argc, char const *argv[]) {
         << "[-c debug_config_file] [-m modbus_mapping] [-g graph_output] [-s maxlogfilesize]\n"
         << "[-mp modbus_port] [-ps persistent_store_port] "
         << "[-cp command/iosh port] [--name device_name] [--stats | --nostats] enable/disable statistics\n"
-        << "[--fix-invalid true [-e ethernet_interface]" << "\n";
+        << "[--fix-invalid true [-e ethernet_interface]"
+        << "\n";
 }
 
 static void listDirectory(const std::string pathToCheck, std::list<std::string> &file_list) {
@@ -811,7 +812,7 @@ void semantic_analysis() {
                      (sm_name == "MQTTBROKER" && (num_sm_params == 2 || num_sm_params == 4)) ||
                      (sm_name == "COUNTERRATE" && (num_sm_params == 3 || num_sm_params == 1))
 #ifdef USE_SDO
-                || (sm_name == "SDOENTRY" && (num_sm_params == 4 || num_sm_params == 5))
+                     || (sm_name == "SDOENTRY" && (num_sm_params == 4 || num_sm_params == 5))
 #endif //USE_SDO
             ) {
             }
@@ -864,7 +865,8 @@ void semantic_analysis() {
                     else {
                         std::stringstream ss;
                         ss << "Error: no instance " << p_i.sValue << " ("
-                           << mi->parameters[i].real_name << ")" << " found for " << mi->getName();
+                           << mi->parameters[i].real_name << ")"
+                           << " found for " << mi->getName();
                         error_messages.push_back(ss.str());
                         MessageLog::instance()->add(ss.str().c_str());
                         ++num_errors;
@@ -1231,7 +1233,7 @@ int loadOptions(int argc, const char *argv[], std::list<std::string> &files) {
 int loadConfig(std::list<std::string> &files) {
     struct timeval now_tv;
     tzset(); /* this initialises the tz info required by ctime().  */
-    srandom(microsecs());
+    srandom(static_cast<unsigned int>(microsecs()));
 
     if (!cw_framework_initialised) {
         int modbus_result = load_preset_modbus_mappings();
